requirements.txt
allure-pytest
allure-python-commons
configparser
faker
flaky
gevent
jose
oauthlib
prospector
py
pytest
pytest-repeat
pytest-rerunfailures
pytest-xdist
python-jose[cryptography]
pytz
requests
requests_oauthlib
xmltodict